extract from a directory
